Clean up
- Don't derive from utility class ToolingApiTestCommon
- formatting

- #19340
