+++ subprojects/tooling-api/src/integTest/groovy/org/gradle/integtests/tooling/ToolingApiIntegrationTest.groovy
@@ -16,23 +16,52 @@
 package org.gradle.integtests.tooling
 
 import org.gradle.api.logging.LogLevel
+import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.RepoScriptBlockUtil
+import org.gradle.integtests.fixtures.executer.GradleDistribution
 import org.gradle.integtests.fixtures.executer.GradleHandle
+import org.gradle.integtests.fixtures.versions.ReleasedVersionDistributions
 import org.gradle.integtests.tooling.fixture.TextUtil
-import org.gradle.integtests.tooling.fixture.ToolingApiTestCommon
+import org.gradle.integtests.tooling.fixture.ToolingApi
+import org.gradle.internal.jvm.Jvm
 import org.gradle.internal.time.CountdownTimer
 import org.gradle.internal.time.Time
+import org.gradle.test.fixtures.file.TestFile
 import org.gradle.tooling.GradleConnector
 import org.gradle.tooling.ProjectConnection
 import org.gradle.tooling.model.GradleProject
 import org.gradle.util.GradleVersion
+import org.junit.Assume
 import spock.lang.Issue
 
 import java.time.ZoneOffset
 import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
 
-class ToolingApiIntegrationTest extends ToolingApiTestCommon {
+import static org.gradle.integtests.tooling.fixture.ToolingApiTestCommon.LOG_LEVEL_TEST_SCRIPT
+import static org.gradle.integtests.tooling.fixture.ToolingApiTestCommon.runLogScript
+import static org.gradle.integtests.tooling.fixture.ToolingApiTestCommon.validateLogs
+
+class ToolingApiIntegrationTest extends AbstractIntegrationSpec {
+
+    final ToolingApi toolingApi = new ToolingApi(distribution, temporaryFolder)
+    final GradleDistribution otherVersion = new ReleasedVersionDistributions().mostRecentRelease
+
+    TestFile projectDir
+//
+    def setup() {
+        projectDir = temporaryFolder.testDirectory
+        // When adding support for a new JDK version, the previous release might not work with it yet.
+        Assume.assumeTrue(otherVersion.worksWith(Jvm.current()))
+
+        settingsFile.touch()
+    }
+
+    void setupLoggingTest() {
+        propertiesFile << "org.gradle.logging.level=quiet"
+        buildFile << LOG_LEVEL_TEST_SCRIPT
+    }
+
 
     def "tooling api uses to the current version of gradle when none has been specified"() {
         projectDir.file('build.gradle') << "assert gradle.gradleVersion == '${GradleVersion.current().version}'"
@@ -87,7 +116,7 @@ class ToolingApiIntegrationTest extends ToolingApiTestCommon {
         when:
         def stdOut = runLogScript(toolingApi, arguments)
         then:
-        true || validateLogs(stdOut, expectedLevel)
+        validateLogs(stdOut, expectedLevel)
 
         where:
         expectedLevel  | arguments
@@ -98,15 +127,14 @@ class ToolingApiIntegrationTest extends ToolingApiTestCommon {
 
     def "tooling api uses the wrapper properties to determine which version to use"() {
         projectDir.file('build.gradle').text = """
-wrapper {
-    distributionUrl = '${otherVersion.binDistribution.toURI()}'
-}
-task check {
-    doLast {
-        assert gradle.gradleVersion == '${otherVersion.version.version}'
-    }
-}
-"""
+        wrapper {
+            distributionUrl = '${otherVersion.binDistribution.toURI()}'
+        }
+        task check {
+            doLast {
+                assert gradle.gradleVersion == '${otherVersion.version.version}'
+            }
+        }"""
         executer.withTasks('wrapper').run()
 
         when:
@@ -122,11 +150,11 @@ task check {
     def "tooling api searches up from the project directory to find the wrapper properties"() {
         projectDir.file('settings.gradle') << "include 'child'"
         projectDir.file('build.gradle') << """
-wrapper { distributionUrl = '${otherVersion.binDistribution.toURI()}' }
-allprojects {
-    task check { doLast { assert gradle.gradleVersion == '${otherVersion.version.version}' } }
-}
-"""
+        wrapper { distributionUrl = '${otherVersion.binDistribution.toURI()}' }
+        allprojects {
+            task check { doLast { assert gradle.gradleVersion == '${otherVersion.version.version}' } }
+        }
+        """
         projectDir.file('child').createDir()
         executer.withTasks('wrapper').run()
